Reduce load factor
Seems like filling the table to only 50% avoids large clusters of filled cells, and hence also long seeks for entries.

Added quite a lot of tests, but this doesn't seem to be a huge cause of delays. Going with it anyway as it seems like a good idea and standard advice for open addressing hash tables.